Delete now works with node deletion

+++ src/main/java/fend/job/job1/definitions/volume/VolumeListController.java
@@ -26,8 +26,9 @@ import javafx.stage.DirectoryChooser;
 import fend.job.job0.JobType0Model;
 import fend.volume.volume0.Volume0;
 import fend.volume.volume1.Volume1;
-import fend.volume.volume2.Volume2;
+import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
@@ -87,6 +88,7 @@ public class VolumeListController {
             volume1.setDbVolume(vol);
             volume1.deleteProperty().addListener(VOLUME_DELETE_LISTENER);
             model.addToVolumeList(volume1);
+            //parentjob.addVolume(volume1);
         }
        
     }
@@ -116,6 +118,7 @@ public class VolumeListController {
                 fevol.setVolume(volumeOnDisk);
                 fevol.deleteProperty().addListener(VOLUME_DELETE_LISTENER);
                 model.addToVolumeList(fevol);
+                //parentjob.addVolume(fevol);
             }
           
         }
@@ -133,36 +136,48 @@ public class VolumeListController {
     private ChangeListener<Boolean> VOLUME_DELETE_LISTENER=new ChangeListener<Boolean>() {
         @Override
         public void changed(ObservableValue<? extends Boolean> observable, Boolean oldValue, Boolean newValue) {
+        if(newValue){
             
-            for(Volume0 vols:model.observableListOfVolumes){
-                boolean volIsToBeDeleted=vols.deleteProperty().get();
-                if(volIsToBeDeleted){
-                    System.out.println(" "+vols.getId()+" : "+vols.getName()+" will be deleted  from the irdb database");
-                    
-                    /**
-                     * delete headers
-                     * delete logs
-                     * delete workflows
-                     **/
-                    Volume dbVol=vols.getDbVolume();
-                    
-                     System.out.println("deleting associated logs");
-                    logService.deleteLogsFor(dbVol);
-                    
-                    System.out.println("deleting associated headers");
-                    headerService.deleteHeadersFor(dbVol);
-                    
-                    
-                    System.out.println("deleting associated workflows");
-                    workflowService.deleteWorkFlowsFor(dbVol);
-                    
-                    System.out.println("deleting volume  from the irdb database");
-                    volumeService.deleteVolume(dbVol.getId());
-                    
-                    
-                    parentjob.removeVolume(vols); 
+                List<Volume0> volTobeParentDeepCopy=new ArrayList<>();
+                for(Volume0 vols:parentjob.getVolumes()){
+                    volTobeParentDeepCopy.add(vols);
                 }
-            }
+                
+                for(Volume0 vols:volTobeParentDeepCopy){
+                    boolean volIsToBeDeleted=vols.deleteProperty().get();
+                    if(volIsToBeDeleted){
+                        System.out.println(" "+vols.getId()+" : "+vols.getName()+" will be deleted  from the irdb database");
+                        
+                        /**
+                         * 
+                         * delete logs
+                         * delete headers
+                         * delete workflows
+                         **/
+                         Volume dbVol=vols.getDbVolume();
+
+                        System.out.println("deleting associated logs");
+                        logService.deleteLogsFor(dbVol);
+
+                        System.out.println("deleting associated headers");
+                        headerService.deleteHeadersFor(dbVol);
+
+
+                        System.out.println("deleting associated workflows");
+                        workflowService.deleteWorkFlowsFor(dbVol);
+
+                        System.out.println("deleting volume  from the irdb database");
+                        volumeService.deleteVolume(dbVol.getId());
+
+                        
+                        
+
+                       
+                        
+                    }
+                }
+        }
+            
         }
     };
